+++ R/class-bal_accuracy.R
@@ -20,6 +20,7 @@ bal_accuracy <- function(data, ...) {
 }
 
 class(bal_accuracy) <- c("class_metric", "function")
+attr(bal_accuracy, "direction") <- "maximize"
 
 #' @export
 #' @rdname bal_accuracy