+++ R/class-f_meas.R
@@ -43,6 +43,7 @@ f_meas <- function(data, ...) {
 }
 
 class(f_meas) <- c("class_metric", "function")
+attr(f_meas, "direction") <- "maximize"
 
 #' @rdname f_meas
 #' @export